Change ui-sref in sidebar to include fragments
This improves navigation when we change state. It allows us to see one
specific part of the website and avoid a strange behavior where we're
stuck at the current scroll level and then miss the begining of a page.

Also change jQuery usage to angular.element to avoid confusion & linting
errors