Add tile design
This commit is kinda huge and change greatly the "offer" page with a new
tile design. It was made generic to be used on other parts of the site
to have consistency.

This also change the bh-yellow color (!!! which need to be renamed !!!)
into a greenish cyan taken from material design UI colors.